reorganized cifar10 example
