--- cleverhans/utils_cifar10.py
@@ -1,42 +1,0 @@
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-from __future__ import unicode_literals
-
-import keras
-from keras.datasets import cifar10
-from keras.utils import np_utils
-
-
-def data_cifar10():
-    """
-    Preprocess CIFAR10 dataset
-    :return:
-    """
-
-    # These values are specific to CIFAR10
-    img_rows = 32
-    img_cols = 32
-    nb_classes = 10
-
-    # the data, shuffled and split between train and test sets
-    (X_train, y_train), (X_test, y_test) = cifar10.load_data()
-
-    if keras.backend.image_dim_ordering() == 'th':
-        X_train = X_train.reshape(X_train.shape[0], 3, img_rows, img_cols)
-        X_test = X_test.reshape(X_test.shape[0], 3, img_rows, img_cols)
-    else:
-        X_train = X_train.reshape(X_train.shape[0], img_rows, img_cols, 3)
-        X_test = X_test.reshape(X_test.shape[0], img_rows, img_cols, 3)
-    X_train = X_train.astype('float32')
-    X_test = X_test.astype('float32')
-    X_train /= 255
-    X_test /= 255
-    print('X_train shape:', X_train.shape)
-    print(X_train.shape[0], 'train samples')
-    print(X_test.shape[0], 'test samples')
-
-    # convert class vectors to binary class matrices
-    Y_train = np_utils.to_categorical(y_train, nb_classes)
-    Y_test = np_utils.to_categorical(y_test, nb_classes)
-    return X_train, Y_train, X_test, Y_test
